requirements.txt
addict==2.2.0
amqp==2.4.0
astroid==2.1.0
autopep8==1.4.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-filter==2.1.0
django-rest-swagger==2.2.0
django-solo==1.1.3
django-timezone-field==3.0
djangorestframework==3.9.1
ez-setup==0.9
httpie==1.0.2
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mysqlclient==1.4.1
openapi-codec==1.3.2
paho-mqtt==1.4.0
pycodestyle==2.4.0
Pygments==2.3.1
pylint==2.2.2
python-crontab==2.3.6
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
wrapt==1.11.1
